+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Added `status` column to `line_items` database table.
 - Added `status` column to `assignments` database table.
 - Added `groupId` column to `line_items` database table.
+- Added possibility to define (optionally) `status` of Line items during ingestion with default value `enabled` for backward compatibility. 
 
 ### Changed
 - Changed availability logic of line items: From now on it is possible to set date restrictions by specifying only starting or ending date.